Initial ETS Rewrite (#497)
* unit test init

* temp

* unit test

* temp
